// spa/angular/src/app/templates/components/funding/interfaces/trading-account.ts
export interface TradingAccount {
  tradingAccountNumber: string;
  typeOfAccount?: string;
  accountName?: string;
  uniqueEntityNumber?: string;
  entity?: string;
  reference?: string;
  bankName?: string;
  bankAccountNumber?: string;
  bankSwiftCode?: string;
  ttBeneBank?: string;
  ttBeneBankAcNm?: string;
  ttBeneBankAcNo?: string;
  ttBeneSwiftRef?: string;
  ttBeneBankCtry?: string;
  ttSwiftRef?: string;
}